Refactor cgn into a single file Update dependencies
