+++ src/controllers/cgnOperatorSearchController.ts
@@ -23,10 +23,10 @@ import { DiscountBucketCode } from "generated/io-cgn-operator-search-api/Discoun
 import { SearchResult } from "generated/io-cgn-operator-search-api/SearchResult";
 import CgnService from "src/services/cgnService";
 
-import { Merchant } from "../../generated/cgn-operator-search/Merchant";
-import { OfflineMerchants } from "../../generated/cgn-operator-search/OfflineMerchants";
-import { OnlineMerchants } from "../../generated/cgn-operator-search/OnlineMerchants";
-import { PublishedProductCategoriesResult } from "../../generated/cgn-operator-search/PublishedProductCategoriesResult";
+import { Merchant } from "../../generated/cgn/Merchant";
+import { OfflineMerchants } from "../../generated/cgn/OfflineMerchants";
+import { OnlineMerchants } from "../../generated/cgn/OnlineMerchants";
+import { PublishedProductCategoriesResult } from "../../generated/cgn/PublishedProductCategoriesResult";
 import { OfflineMerchantSearchRequest } from "../../generated/io-cgn-operator-search-api/OfflineMerchantSearchRequest";
 import { OnlineMerchantSearchRequest } from "../../generated/io-cgn-operator-search-api/OnlineMerchantSearchRequest";
 import { SearchRequest } from "../../generated/io-cgn-operator-search-api/SearchRequest";
