+++ src/services/cgnService.ts
@@ -2,12 +2,12 @@
  * This service interactsnwith the Bonus API
  */
 
-import { Card } from "@pagopa/io-functions-cgn-sdk/Card";
-import { CgnActivationDetail } from "@pagopa/io-functions-cgn-sdk/CgnActivationDetail";
-import { EycaActivationDetail } from "@pagopa/io-functions-cgn-sdk/EycaActivationDetail";
-import { EycaCard } from "@pagopa/io-functions-cgn-sdk/EycaCard";
-import { InstanceId } from "@pagopa/io-functions-cgn-sdk/InstanceId";
-import { Otp } from "@pagopa/io-functions-cgn-sdk/Otp";
+import { Card } from "../../generated/cgn/Card";
+import { CgnActivationDetail } from "../../generated/cgn/CgnActivationDetail";
+import { EycaActivationDetail } from "../../generated/cgn/EycaActivationDetail";
+import { EycaCard } from "../../generated/cgn/EycaCard";
+import { InstanceId } from "../../generated/cgn/InstanceId";
+import { Otp } from "../../generated/cgn/Otp";
 import {
   IResponseErrorConflict,
   IResponseErrorForbiddenNotAuthorized,
@@ -17,6 +17,7 @@ import {
   IResponseSuccessAccepted,
   IResponseSuccessJson,
   IResponseSuccessRedirectToResource,
+  ProblemJson,
   ResponseErrorConflict,
   ResponseErrorForbiddenNotAuthorized,
   ResponseErrorInternal,
@@ -66,7 +67,7 @@ export default class CgnService {
           case 404:
             return ResponseErrorNotFound("Not Found", "CGN not found");
           case 500:
-            return ResponseErrorInternal(readableProblem(response.value));
+            return ResponseErrorInternal(readableProblem(response.value as ProblemJson));
           default:
             return ResponseErrorStatusNotDefinedInSpec(response);
         }
@@ -106,7 +107,7 @@ export default class CgnService {
               "EYCA Card is missing while citizen is eligible to obtain it or a CGN is already activated"
             );
           case 500:
-            return ResponseErrorInternal(readableProblem(response.value));
+            return ResponseErrorInternal(readableProblem(response.value as ProblemJson));
           default:
             return ResponseErrorStatusNotDefinedInSpec(response);
         }
@@ -154,7 +155,7 @@ export default class CgnService {
               "Cannot start a new CGN activation because the CGN is already active, revoked or expired"
             );
           case 500:
-            return ResponseErrorInternal(readableProblem(response.value));
+            return ResponseErrorInternal(readableProblem(response.value as ProblemJson));
           default:
             return ResponseErrorStatusNotDefinedInSpec(response);
         }
@@ -189,7 +190,7 @@ export default class CgnService {
               "No User CGN activation found"
             );
           case 500:
-            return ResponseErrorInternal(readableProblem(response.value));
+            return ResponseErrorInternal(readableProblem(response.value as ProblemJson));
           default:
             return ResponseErrorStatusNotDefinedInSpec(response);
         }
@@ -237,7 +238,7 @@ export default class CgnService {
               "Cannot start a new EYCA activation because EYCA card is already active, revoked or expired"
             );
           case 500:
-            return ResponseErrorInternal(readableProblem(response.value));
+            return ResponseErrorInternal(readableProblem(response.value as ProblemJson));
           default:
             return ResponseErrorStatusNotDefinedInSpec(response);
         }
@@ -272,7 +273,7 @@ export default class CgnService {
               "No EYCA Card activation found"
             );
           case 500:
-            return ResponseErrorInternal(readableProblem(response.value));
+            return ResponseErrorInternal(readableProblem(response.value as ProblemJson));
           default:
             return ResponseErrorStatusNotDefinedInSpec(response);
         }
@@ -320,7 +321,7 @@ export default class CgnService {
               "Cannot start a new CGN unsubscription"
             );
           case 500:
-            return ResponseErrorInternal(readableProblem(response.value));
+            return ResponseErrorInternal(readableProblem(response.value as ProblemJson));
           default:
             return ResponseErrorStatusNotDefinedInSpec(response);
         }
@@ -352,7 +353,7 @@ export default class CgnService {
           case 403:
             return ResponseErrorForbiddenNotAuthorized;
           case 500:
-            return ResponseErrorInternal(readableProblem(response.value));
+            return ResponseErrorInternal(readableProblem(response.value as ProblemJson));
           default:
             return ResponseErrorStatusNotDefinedInSpec(response);
         }
